Adds `smarty` internal package
- Houses logic specific to smarty api interactions into a dedicated internal package.
- Updates go version in go.mod from 1.17 to 1.21 in order utilize go's built-in min function.

Tested:
- Ran each command in cmd to ensure they were still working as expected with Go's version update.
- Ran the smarty-usps-comparison tool to ensure discrepancies between USPS and Smarty data were still being appropriately logged.
