+++ cmd/smarty-usps-comparison/main.go
@@ -1,117 +1,24 @@
 package main
 
 import (
-	"bytes"
 	"encoding/csv"
-	"encoding/json"
 	"flag"
 	"fmt"
 	"io"
 	"log"
-	"net/http"
 	"os"
 	"sort"
 	"strings"
 	"time"
+
+	"github.com/corbaltcode/usps/internal/smarty"
 )
-
-type SmartyResponse struct {
-	InputIndex int       `json:"input_index"`
-	Status     string    `json:"status,omitempty"`
-	Reason     string    `json:"reason,omitempty"`
-	Zipcodes   []Zipcode `json:"zipcodes,omitempty"`
-}
-
-type Zipcode struct {
-	Zipcode           string            `json:"zipcode"`
-	ZipcodeType       string            `json:"zipcode_type"`
-	DefaultCity       string            `json:"default_city"`
-	CountyFIPS        string            `json:"county_fips"`
-	CountyName        string            `json:"county_name"`
-	StateAbbreviation string            `json:"state_abbreviation"`
-	State             string            `json:"state"`
-	Latitude          float64           `json:"latitude"`
-	Longitude         float64           `json:"longitude"`
-	Precision         string            `json:"precision"`
-	AlternateCounties []AlternateCounty `json:"alternate_counties"`
-}
-
-type AlternateCounty struct {
-	CountyFIPS        string `json:"county_fips"`
-	CountyName        string `json:"county_name"`
-	StateAbbreviation string `json:"state_abbreviation"`
-	State             string `json:"state"`
-}
 
 type ZipcodeResult struct {
 	Zipcode      string
 	USPSFips     []string
 	SmartyFips   []string
 	ErrorMessage string
-}
-
-type ZipcodeRequest struct {
-	Zipcode string `json:"zipcode"`
-}
-
-type SmartyClient struct {
-	AuthId    string
-	AuthToken string
-	BaseURL   string
-}
-
-func NewSmartyClient(authId, authToken string) *SmartyClient {
-	return &SmartyClient{
-		AuthId:    authId,
-		AuthToken: authToken,
-		BaseURL:   "https://us-zipcode.api.smarty.com/lookup",
-	}
-}
-
-func (client *SmartyClient) QueryBatch(zips []string) ([]SmartyResponse, error) {
-	var payload []ZipcodeRequest
-	for _, zip := range zips {
-		payload = append(payload, ZipcodeRequest{Zipcode: zip})
-	}
-	jsonPayload, err := json.Marshal(payload)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal JSON payload: %w", err)
-	}
-
-	apiURL := fmt.Sprintf("%s?auth-id=%s&auth-token=%s", client.BaseURL, client.AuthId, client.AuthToken)
-
-	req, err := http.NewRequest("POST", apiURL, bytes.NewBuffer(jsonPayload))
-	if err != nil {
-		return nil, fmt.Errorf("failed to create HTTP request: %w", err)
-	}
-	req.Header.Set("Content-Type", "application/json; charset=utf-8")
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("failed to query Smarty API: %w", err)
-	}
-	defer resp.Body.Close()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read response body: %w", err)
-	}
-
-	if resp.StatusCode == http.StatusTooManyRequests {
-		return nil, fmt.Errorf("rate limit exceeded: %s", body)
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("API request failed with status %d: %s", resp.StatusCode, body)
-	}
-
-	var smartyResponses []SmartyResponse
-	err = json.Unmarshal(body, &smartyResponses)
-	if err != nil {
-		return nil, fmt.Errorf("failed to unmarshal response: %w", err)
-	}
-
-	return smartyResponses, nil
 }
 
 func parseCSVList(dataString string) []string {
@@ -155,7 +62,7 @@ func readZipToCountyDataFromCSV(reader io.Reader) ([]string, map[string][]string
 	return zips, zipToCounty, nil
 }
 
-func ProcessSmartyResponse(responseBody []SmartyResponse, zipToCounties map[string][]string, zipcodes []string, startIndex int) ([]ZipcodeResult, error) {
+func ProcessSmartyResponse(responseBody []smarty.SmartyResponse, zipToCounties map[string][]string, zipcodes []string, startIndex int) ([]ZipcodeResult, error) {
 	processedResponses := make([]ZipcodeResult, 0)
 
 	for _, response := range responseBody {
@@ -255,7 +162,7 @@ func setupCSVWriter(output io.Writer) *csv.Writer {
 func main() {
 	authId := mustGetenv("AUTH_ID")
 	authToken := mustGetenv("AUTH_TOKEN")
-	client := NewSmartyClient(authId, authToken)
+	client := smarty.NewSmartyClient(authId, authToken)
 	uspsZipToCountyFile := flag.String("csv", "", "CSV file path containing zip to county mappings for USPS data")
 	flag.Parse()
